Recognize Android motion photo videos
Fix handling of motion photo videos: do not count them as missing.

Thanks @mraedis for reporting this issue.

In the database, motion photo videos are tracked in the `asset` table.
On disk, these extracted video parts are saved to the `encoded-video`
directory with `MP` suffix and `.mp4` extension.
Now the script checks for these files on disk.